requirements.txt
bleach==1.5.0
boost==0.1
certifi==2017.11.5
chardet==3.0.4
click==6.7
decorator==4.2.1
dlib==19.9.0
enum34==1.1.6
face-recognition==1.0.0
face-recognition-models==0.3.0
Flask==0.12.2
html5lib==0.9999999
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
Mastodon.py==1.2.1
numpy==1.14.0
opencv-python==3.4.0.12
Pillow==5.0.0
protobuf==3.5.1
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
scipy==1.0.0
six==1.11.0
SQLAlchemy==1.2.2
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0
urllib3==1.22
Werkzeug==0.14.1
wincertstore==0.2
